size t

+++ cpp/src/barretenberg/stdlib/commitment/bench_bindings.hpp
@@ -9,6 +9,6 @@ std::shared_ptr<barretenberg::srs::factories::CrsFactory> create_prover_factory(
 void build_circuit(UltraPlonkComposer& composer, size_t circuit_size);
 
 extern "C" {
-UltraPlonkComposer* create_composer(int circuit_size);
+UltraPlonkComposer* create_composer(size_t circuit_size);
 void commit(UltraPlonkComposer* composer, size_t length);
 }
